renderFinal: evaluate layers bottom-up
This change makes it so that layers are evaluated starting with the
bottom-most layer. Even though we still return the resulting images in
the correct order for Vty (top-most image first), this change results in
the rendering state monad getting updated with the extends and viewport
information resulting from the bottom-most layer first, progressing up
the layers after that.

This change is important because it has the consequence that upper
layers can now obtain rendering metadata (such as extents) to make
decisions such as overlay translated images at the right locations.
While this change does not actually expose any new API to access
rendering metadata, it is a necessary first step to make such metadata
access useful.